MyStudyZone/requirements.txt
altgraph==0.16.1
beautifulsoup4==4.6.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
diff-match-patch==20121119
Django==2.0.1
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.12
django-simple-captcha==0.5.9
easygui==0.98.1
et-xmlfile==1.0.1
Faker==0.9.2
Flask==1.0.2
future==0.16.0
httplib2==0.11.3
idna==2.7
itchat==1.3.10
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
lxml==4.2.4
macholib==1.11
MarkupSafe==1.0
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.4.9
pefile==2018.8.8
Pillow==5.3.0
pyHook==1.5.1
pypng==0.0.18
PyQRCode==1.2.1
pytesseract==0.2.5
python-dateutil==2.7.3
pytz==2017.3
PyUserInput==0.1.11
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==3.12
requests==2.19.1
schedule==0.5.0
selenium==3.14.0
six==1.11.0
splinter==0.8.0
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
urllib3==1.23
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
